Handle similar gem names in updated_version_for
Bummr can get confused when updating a gem when there is another gem
used by the project that is named similarly but with a prefix. E.g.
when updating websockets, bummr can get confused by the presence of
em-websockets and pick the version number of em-websockets for the
commit message.

This commit fixes that problem by anchoring the regex on the directory
separator.

Note that I'm not sure how this behaves on Windows. My hope is that
`bundle list --paths` uses the directory separator as is specified in
`File::SEPARATOR`, but this is untested.